mkimage: add support for rpi4-based LXD host
this is not intended for use as a router, but for a generic LXD host
that acts as a DHCP client inside an existing lan.

I figured out that by installing `mount` and using a GPT partition table
I can use partlabels instead of hardcoding partitions. I applied the
requires changes globally in the hope that I'll be able to do the same
for my homeserver.