bitwarden: update to 2025.7.0-beta
The service now fails to start if the installation ID is all zeroes.
Since this instance does not have internet access and thus no way to
check if the ID is valid, I simply changed it to something that's not 0
which makes it work again.

See https://github.com/bitwarden/self-host/issues/361